Review `ruff_source_file`, merge in our line ending tools
