+++ Cargo.toml
@@ -29,7 +29,6 @@ air_r_syntax = { path = "./crates/air_r_syntax" }
 biome_ungrammar = { path = "./crates/biome_ungrammar" }
 fs = { path = "./crates/fs" }
 line_ending = { path = "./crates/line_ending" }
-lsp = { path = "./crates/lsp" }
 lsp_test = { path = "./crates/lsp_test" }
 tests_macros = { path = "./crates/tests_macros" }
 ruff_server = { path = "./crates/ruff_server" }
