+++ crates/air_r_formatter/Cargo.toml
@@ -24,7 +24,6 @@ tracing = { workspace = true }
 air_formatter_test = { workspace = true }
 air_r_parser = { workspace = true }
 biome_parser = { workspace = true }
-line_ending = { workspace = true }
 tests_macros = { workspace = true }
 
 [lints]
